Extract JSON_OBJECT() and JSON_ARRAY() into own class
