+++ h2/src/test/org/h2/test/db/TestAlterTableNotFound.java
@@ -8,7 +8,6 @@ package org.h2.test.db;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
-import org.h2.jdbc.JdbcConnection;
 import org.h2.test.TestBase;
 import org.h2.test.TestDb;
